# api/openapi.yaml
openapi: 3.1.0
info:
  title: n0passtemps
  version: 1.1.0
  summary: On-premise passwordless authentication server.
  description: |
    n0passtemps authenticates the end users of an integrating application with
    WebAuthn, TOTP and single-use recovery codes. It is deployed on premise and
    is called server to server: the browser talks to the integrating
    application, and the integrating application talks to this service with an
    API key.

    # Two authenticated surfaces

    `/v1` is the surface an integrating application calls. Every route on it
    requires an API key. It is not anonymous: an unauthenticated registration
    route would let anyone enrol an authenticator against any subject.

    `/admin/v1` is the surface an operator calls. Every route on it requires an
    administrative token, carries a role check, and sits behind a network allow
    list configured on the server. The two credential kinds cannot be
    substituted for each other.

    Three routes carry no credential at all: the liveness probe and the two
    paths that serve the JWKS document. Both disclose only public information.
    A fourth unauthenticated case is not a route in this document: `OPTIONS`
    under `/v1/` is answered by the CORS middleware with `204` and no body,
    because a preflight request carries no credential by definition.

    # The two-round-trip ceremony flow

    Both WebAuthn ceremonies are two round trips, and the caller sits in the
    middle of them.

...

...

...

...

...

...
servers:
...
paths:
...
  /v1/subjects/{subject_ref}:
    parameters:
      - $ref: "#/components/parameters/SubjectRef"
    get:
      tags: [Subjects]
      operationId: getSubject
      summary: Read a subject's enrolled factors.
      description: |
        Requires an API key that holds the `subjects` scope, or an unrestricted
        key.

        Used to decide what to offer a returning user: a subject with no
        credential needs registration, one with credentials needs an assertion.

        Unlike the ceremony routes, this one distinguishes an unknown subject
        with a `404`, because it is a lookup rather than an authentication
        attempt.
      security:
        - apiKey: []
      responses:
        "200":
          description: The subject.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/Throttled"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: |
        The credential of an integrating application, required on every `/v1`
        route except the liveness probe and the two JWKS paths.

        It is presented as `Authorization: Bearer npt_<selector>.<verifier>`,
        for example
        `npt_0000000000000000.AAAAAAAAAAAAAAAAAAAAAAAAAAA` with the zeros
        replaced by the real value. The selector is sixteen lowercase hex
        characters and is stored in clear so that authentication is one indexed
        lookup; the verifier is 160 bits of base64url with no padding and only a
        digest of it is stored. The `npt_` prefix exists so that a leaked token
        is recognisable to a secret scanner.

        A token is shown exactly once, by `POST /admin/v1/api-keys`. There is no
        operation that can show it again; mint a replacement instead.

        A key may be restricted to named route families when it is minted. A
        key calling outside its scopes is answered `403`, and a key minted with
        no scopes reaches every family. See "API key scopes" in the
        introduction.

        A token minted for the administrative surface cannot authenticate here,
        and the reverse: the kind is bound into the stored digest as well as
        checked on the way in.
    adminToken:
      type: http
      scheme: bearer
      description: |
        The credential of an operator, required on every `/admin/v1` route.

        It is presented as `Authorization: Bearer npa_<selector>.<verifier>`,
        with the same `selector.verifier` shape as an API key and the `npa_`
        prefix in place of `npt_`. Each token carries one of three roles, and
        each route additionally requires a named permission.

        A token is shown exactly once, by `POST /admin/v1/admin-tokens`, and
        cannot be retrieved again.
...
  headers:
    XRequestId:
      description: |
        The request identifier, present on every response including error
        responses. A client-supplied `X-Request-Id` is echoed when it is at most
        64 characters of letters, digits, hyphen, underscore and full stop;
        anything else is replaced by a value the service assigns. Quote it in a
        support exchange.
      schema:
        type: string
        maxLength: 64
    RetryAfter:
      description: |
        How many seconds to wait before retrying. Present on a `429`, and
        mirrored in the problem body as `retry_after_seconds`.
      schema:
        type: integer
        minimum: 1
    WWWAuthenticate:
      description: |
        The challenge RFC 9110 section 11.6.1 requires on a `401`. Fixed at
        `Bearer realm="n0passtemps"`.
      schema:
        type: string
  parameters:
    SubjectRef:
      name: subject_ref
      in: path
      required: true
      description: |
        The identifier the calling application uses for its user. It is opaque
        to this service, must not be empty and must contain no control
        characters or invalid UTF-8, and is bounded by
        `subject.max_ref_length`, 256 bytes by default.

        Use an opaque identifier rather than an email address. This service logs
        the route pattern rather than the concrete path, so the value does not
        reach its own logs, but an intervening reverse proxy will log it unless
        configured otherwise. The reference is encrypted at rest regardless.
      schema:
        type: string
        minLength: 1
        maxLength: 256
...
  responses:
    BadRequest:
      description: |
        Malformed or invalid input. This is the one class of failure where
        `detail` is populated, because telling a caller that a field is missing
        helps them and tells an attacker nothing.

        Unknown JSON fields are refused rather than ignored, so a misspelled
        field name is reported instead of silently taking its zero value. A body
        containing more than one JSON document is refused as well.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            missingField:
              value:
                type: urn:n0passtemps:error:bad-request
                title: the request is not valid
                status: 400
                detail: challenge_id is required
                request_id: 9f2a1c7b5d3e4f60a1b2c3d4
    Unauthorized:
      description: |
        The bearer credential is missing, malformed, unrecognised, expired or
        revoked. The five cases are one response on purpose: distinguishing them
        would let a caller enumerate valid selectors.

        A credential rejection is audited and raises an alert, because a run of
        them is the earliest visible sign of a leaked key being probed.

        A credential store that cannot be reached is not this response. It is a
        `503`, because nothing was rejected.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        WWW-Authenticate:
          $ref: "#/components/headers/WWWAuthenticate"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            noCredential:
              value:
                type: urn:n0passtemps:error:unauthorized
                title: authentication is required
                status: 401
                request_id: 9f2a1c7b5d3e4f60a1b2c3d4
...
    Forbidden:
      description: |
        The credential is valid but is not permitted to perform the operation.

        On `/v1` this is an API key calling a route outside its scopes. The
        attempt is audited as `api_key.rejected` with the route and the scopes
        the key holds. Re-authenticating does not help; the key has to be
        replaced by one minted with the scope the route needs.

        On `/admin/v1` this is a role that does not hold the route's permission,
        an attempt to decide one's own approval request, or a refused
        `X-Approval-Id` redemption. Every redemption refusal is this one
        response, whether the approval is unknown, names another operation,
        belongs to another requester, is not in the approved state, has
        expired, was already spent, or covers a different payload; the reason
        is audited as `approval.rejected` and not returned. On either surface it
        is also the answer when the credential belongs to a tenant this
        deployment does not serve. On a WebAuthn registration it can mean the
        authenticator model is not permitted by the configured AAGUID policy,
        which is the one forbidden case that is disclosed, because a user
        holding an unsupported key needs to be told to use a different one.

        A permission denial is audited and raises an alert. The reason names the
        role and the permission, so it is logged rather than returned.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            wrongRole:
              value:
                type: urn:n0passtemps:error:forbidden
                title: this credential is not permitted to perform that operation
                status: 403
                request_id: 9f2a1c7b5d3e4f60a1b2c3d4
    NotFound:
      description: |
        The resource does not exist. It is also the answer when a resource
        exists but under a different parent, so that a route cannot be used to
        discover ownership, and when no route matches the request at all.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            absent:
              value:
                type: urn:n0passtemps:error:not-found
                title: the resource does not exist
                status: 404
                request_id: 9f2a1c7b5d3e4f60a1b2c3d4
...
    Throttled:
      description: |
        A rate limit was reached. Limits apply per subject, per end-user address
        and per API key at the same time, and the two administrative revoke
        routes additionally share one per administrator. The response does not
        say which dimension tripped.

        The per-subject limit is held once per factor. Failures at TOTP lock
        the subject out of TOTP and leave WebAuthn and recovery codes usable,
        and the reverse, so nobody can be kept from their passkey by somebody
        else sending wrong codes in their name. The per-address limit works
        from the `X-End-User-IP` request header and is not applied without it.
        Both lock out for `throttle.lockout_duration`.

        The per-key limit is a rate limit and not a lockout: `Retry-After`
        runs to the end of the current `throttle.window`, after which the key
        is served again.

        The per-key limit, `throttle.max_requests_per_key`, is metered on every
        `/v1` route that takes an API key, including the ones that perform no
        authentication: resolving a subject, starting a ceremony and reading
        the health report all count. A leaked key cannot be used without bound
        on any of them.

        The check runs before any expensive work, so a locked-out caller does
        not get a signature verification or an Argon2id evaluation out of the
        service.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            throttled:
              value:
                type: urn:n0passtemps:error:throttled
                title: too many attempts
                status: 429
                retry_after_seconds: 900
                request_id: 9f2a1c7b5d3e4f60a1b2c3d4
    Unavailable:
      description: |
        The credential store could not be reached, so the bearer credential
        could not be checked. This is the service's fault rather than the
        caller's: the credential was neither accepted nor rejected, nothing is
        audited as a rejection, and no alert is raised. Retry with the same
        credential. The liveness probe reports the same condition.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            storeDown:
              value:
                type: urn:n0passtemps:error:unavailable
                title: the service is temporarily unavailable
                status: 503
                request_id: 9f2a1c7b5d3e4f60a1b2c3d4
...
    InternalError:
      description: |
        A fault in the service. Nothing about the cause is disclosed; it is
        logged at error level against the same `request_id`.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            internal:
              value:
                type: urn:n0passtemps:error:internal
                title: the service could not complete the request
                status: 500
                request_id: 9f2a1c7b5d3e4f60a1b2c3d4
  schemas:
    Problem:
      type: object
      title: Problem
      description: |
        The RFC 9457 error body returned by every refusal, with the media type
        `application/problem+json`.

        The members are deliberately spare. `title` and `type` describe a class
        of failure, never the specific reason one request was refused: an
        authentication service that explains why a ceremony failed is an oracle
        for probing its own configuration. The real reason is logged and
        audited, where an operator can read it and an attacker cannot.
      required:
        - type
        - title
        - status
      properties:
        type:
          $ref: "#/components/schemas/ProblemType"
        title:
          type: string
          description: |
            A short, fixed description of the class of failure. It is stable for
            a given `type` and is not customised per request.
          examples:
            - authentication did not succeed
        status:
          type: integer
          description: |
            The HTTP status code repeated, so a client that has only the body
            still knows.
          examples:
            - 401
        detail:
          type: string
          description: |
            Present only where the extra information is safe to disclose, which
            in practice means input validation and state conflicts. Absent on
            every authentication and authorisation failure.
        request_id:
          type: string
          description: |
            Matches the `X-Request-Id` response header, and is the identifier an
            operator uses to find the corresponding log and audit entries.
          maxLength: 64
        retry_after_seconds:
          type: integer
          description: |
            Set on a throttled response only, mirroring the `Retry-After`
            header. Never below 1.
          minimum: 1
    ProblemType:
      type: string
      title: ProblemType
      description: |
        A stable identifier for a class of error, usable by a client to branch
        on. It is a URN rather than a URL because there is no documentation site
        to dereference.

        The status each one is returned with:

        | type | status |
        | --- | --- |
        | `urn:n0passtemps:error:bad-request` | 400 |
        | `urn:n0passtemps:error:unauthorized` | 401 |
        | `urn:n0passtemps:error:ceremony-failed` | 401 |
        | `urn:n0passtemps:error:forbidden` | 403 |
        | `urn:n0passtemps:error:not-found` | 404 |
        | `urn:n0passtemps:error:conflict` | 409 |
        | `urn:n0passtemps:error:payload-too-large` | 413 |
        | `urn:n0passtemps:error:unsupported-media-type` | 415 |
        | `urn:n0passtemps:error:throttled` | 429 |
        | `urn:n0passtemps:error:internal` | 500 |
        | `urn:n0passtemps:error:unavailable` | 503 |
        | `urn:n0passtemps:error:approval-required` | 202 |

        `approval-required` is the one that is not a failure: it reports that a
        sensitive operation was queued for a second administrator rather than
        performed, and it is carried on a `202` with a problem body.

        `unavailable` is returned by every authenticated route when the
        credential store cannot be reached. The liveness probe reports the same
        condition with a `503` and a liveness body rather than a problem
        document.
      enum:
        - urn:n0passtemps:error:bad-request
        - urn:n0passtemps:error:unauthorized
        - urn:n0passtemps:error:forbidden
        - urn:n0passtemps:error:not-found
        - urn:n0passtemps:error:conflict
        - urn:n0passtemps:error:ceremony-failed
        - urn:n0passtemps:error:throttled
        - urn:n0passtemps:error:payload-too-large
        - urn:n0passtemps:error:unsupported-media-type
        - urn:n0passtemps:error:approval-required
        - urn:n0passtemps:error:internal
        - urn:n0passtemps:error:unavailable
...
    SubjectStatus:
      type: string
      title: SubjectStatus
      enum:
        - active
        - locked
        - pending_deletion
...
    SubjectSummary:
      type: object
      title: SubjectSummary
      description: |
        What the public routes disclose about a subject. It carries the internal
        identifier and the status, and never the reference the caller supplied.
      required:
        - subject_id
        - status
        - created_at
        - credential_count
        - totp_enrolled
        - recovery_codes_remaining
      properties:
        subject_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/SubjectStatus"
        created_at:
          type: string
          format: date-time
        credential_count:
          type: integer
          description: |
            Active WebAuthn credentials. Revoked ones are not counted here; the
            administrative routes include them.
        totp_enrolled:
          type: boolean
          description: |
            True only once an enrolment has been confirmed. A pending secret
            does not count.
        recovery_codes_remaining:
          type: integer

    # The two ceremony payloads below are passed through verbatim in both
    # directions. They are declared as open objects on purpose: modelling the
    # whole of the W3C Web Authentication schema inline would duplicate a
    # specification that changes independently of this one, and a schema that
    # fell behind it would start rejecting valid browser output. The
    # authoritative definitions are PublicKeyCredentialCreationOptions,
    # PublicKeyCredentialRequestOptions and PublicKeyCredential in the Web
    # Authentication Level 2 specification.
...